requirements.txt
torch>=1.9.0
transformers>=4.12.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
tensorboard>=2.7.0
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
jupyter>=1.0.0
notebook>=6.4.0